ML_Backend/requirements.txt
# requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
numpy>=1.26.0
opencv-python>=4.8.0.74
aiohttp==3.8.5
openai==0.28.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
easyocr==1.7.0
groq
requests==2.31.0